backend/services/menu-service/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
pydantic